📑 :: 유저 엔티티 수정
