+++ src/main/java/com/example/soopgwan/domain/user/domain/User.java
@@ -1,0 +1,38 @@
+package com.example.soopgwan.domain.user.domain;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import javax.persistence.*;
+
+@Entity
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+@Table(name = "tbl_user")
+public class User {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(columnDefinition = "BIGINT", nullable = false)
+    private long id;
+
+    @Column(columnDefinition = "VARCHAR(24)", nullable = false)
+    private String accountId;
+
+    @Column(columnDefinition = "CHAR(60)", nullable = false)
+    private String password;
+
+    @Column(columnDefinition = "VARCHAR(10)", nullable = false)
+    private String name;
+
+    @Column(columnDefinition = "CHAR(11)", nullable = false)
+    private String phoneNumber;
+
+    @Column(columnDefinition = "CHAR(3)", nullable = false)
+    private String level;
+}
+
