--- src/main/java/com/example/soopgwan/entity/User.java
@@ -1,33 +1,0 @@
-package com.example.soopgwan.entity;
-
-import lombok.*;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.Id;
-import javax.persistence.Table;
-
-@Entity
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-@Table(name = "tbl_user")
-public class User {
-    @Id
-    @Column(length = 24, nullable = false)
-    private String accountId;
-
-    @Column(length = 60, nullable = false)
-    private String password;
-
-    @Column(length = 10, nullable = false)
-    private String name;
-
-    @Column(length = 11, nullable = false)
-    private String phoneNumber;
-
-    @Column(length = 3, nullable = false)
-    private String level;
-}
